# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "BBoxAnnotation": {
        "description": "One 2D bounding box on a captured RGB frame (illustrative annotations).",
        "properties": {
          "frame": {
            "title": "Frame",
            "type": "integer"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "x": {
            "title": "X",
            "type": "integer"
          },
          "y": {
            "title": "Y",
            "type": "integer"
          }
        },
        "required": [
          "frame",
          "label",
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BBoxAnnotation",
        "type": "object"
      },
      "DailyFrameCount": {
        "description": "Frames ingested into the lake on a given day (ingest throughput).",
        "properties": {
          "date": {
            "title": "Date",
            "type": "string"
          },
          "frames": {
            "title": "Frames",
            "type": "integer"
          }
        },
        "required": [
          "date",
          "frames"
        ],
        "title": "DailyFrameCount",
        "type": "object"
      },
...
      "EpisodeDetail": {
        "description": "Episode detail: the persisted metadata plus live-computed storage totals.",
        "properties": {
          "metadata": {
            "$ref": "#/components/schemas/EpisodeMetadata"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "size_human": {
            "title": "Size Human",
            "type": "string"
          },
          "total_objects": {
            "title": "Total Objects",
            "type": "integer"
          }
        },
        "required": [
          "metadata",
          "total_objects",
          "size_bytes",
          "size_human"
        ],
        "title": "EpisodeDetail",
        "type": "object"
      },
      "EpisodeMetadata": {
        "description": "The full per-episode record, persisted as episodes/<id>/metadata.json.",
        "properties": {
          "annotations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/BBoxAnnotation"
            },
            "title": "Annotations",
            "type": "array"
          },
          "capture_source": {
            "default": "carla",
            "title": "Capture Source",
            "type": "string"
          },
          "captured_frames": {
            "title": "Captured Frames",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "fps": {
            "title": "Fps",
            "type": "integer"
          },
          "frames_by_sensor": {
            "additionalProperties": {
              "type": "integer"
            },
            "default": {},
            "title": "Frames By Sensor",
            "type": "object"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "requested_frames": {
            "title": "Requested Frames",
            "type": "integer"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "scenario_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Name"
          },
          "sensors": {
            "items": {
              "type": "string"
            },
            "title": "Sensors",
            "type": "array"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "town": {
            "title": "Town",
            "type": "string"
          },
          "traffic_density": {
            "title": "Traffic Density",
            "type": "string"
          },
          "weather": {
            "title": "Weather",
            "type": "string"
          }
        },
        "required": [
          "id",
          "town",
          "weather",
          "traffic_density",
          "fps",
          "requested_frames",
          "captured_frames",
          "sensors",
          "status",
          "started_at"
        ],
        "title": "EpisodeMetadata",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Scenario": {
        "description": "A stored Scenario: the input config plus server-assigned identity.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "fps": {
            "title": "Fps",
            "type": "integer"
          },
          "frame_count": {
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Frame Count",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "sensors": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Sensors",
            "type": "array"
          },
          "town": {
            "title": "Town",
            "type": "string"
          },
          "traffic_density": {
            "title": "Traffic Density",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "weather": {
            "title": "Weather",
            "type": "string"
          }
        },
        "required": [
          "name",
          "town",
          "weather",
          "traffic_density",
          "fps",
          "frame_count",
          "sensors",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "Scenario",
        "type": "object"
      },
      "ScenarioInput": {
        "description": "Create/edit payload for a Scenario (edit fully replaces the config).",
        "properties": {
          "description": {
            "default": "",
            "maxLength": 500,
            "title": "Description",
            "type": "string"
          },
          "fps": {
            "title": "Fps",
            "type": "integer"
          },
          "frame_count": {
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Frame Count",
            "type": "integer"
          },
          "name": {
            "maxLength": 80,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "sensors": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Sensors",
            "type": "array"
          },
          "town": {
            "title": "Town",
            "type": "string"
          },
          "traffic_density": {
            "title": "Traffic Density",
            "type": "string"
          },
          "weather": {
            "title": "Weather",
            "type": "string"
          }
        },
        "required": [
          "name",
          "town",
          "weather",
          "traffic_density",
          "fps",
          "frame_count",
          "sensors"
        ],
        "title": "ScenarioInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Local API for CARLA Sensor Data Lake: manage CARLA scenario configs, drive the simulator, and browse and serve the resulting multi-sensor driving datasets (RGB, segmentation, depth, LiDAR, telemetry) stored in Backblaze B2 over the S3-compatible API. This contract documents the app's local API, not a hosted public endpoint.",
    "title": "CARLA Sensor Data Lake API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/episodes/{episode_id}": {
      "delete": {
        "operationId": "delete_episode_endpoint_episodes__episode_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Episode Endpoint",
        "tags": [
          "episodes"
        ]
      },
      "get": {
        "operationId": "get_episode_endpoint_episodes__episode_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "episode_id",
            "required": true,
            "schema": {
              "title": "Episode Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Episode Endpoint",
        "tags": [
          "episodes"
        ]
      }
    },
...
    "/lake/ingest": {
      "get": {
        "operationId": "lake_ingest_endpoint_lake_ingest_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "title": "Days",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DailyFrameCount"
                  },
                  "title": "Response Lake Ingest Endpoint Lake Ingest Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Lake Ingest Endpoint",
        "tags": [
          "lake"
        ]
      }
    },
...
    "/scenarios": {
      "get": {
        "operationId": "list_scenarios_endpoint_scenarios_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Scenario"
                  },
                  "title": "Response List Scenarios Endpoint Scenarios Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Scenarios Endpoint",
        "tags": [
          "scenarios"
        ]
      },
      "post": {
        "operationId": "create_scenario_endpoint_scenarios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Scenario Endpoint",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/scenarios/{scenario_id}": {
      "delete": {
        "operationId": "delete_scenario_endpoint_scenarios__scenario_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Scenario Endpoint",
        "tags": [
          "scenarios"
        ]
      },
      "get": {
        "operationId": "get_scenario_endpoint_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Scenario Endpoint",
        "tags": [
          "scenarios"
        ]
      },
      "put": {
        "operationId": "update_scenario_endpoint_scenarios__scenario_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Scenario"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Scenario Endpoint",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/scenarios/{scenario_id}/run": {
      "post": {
        "description": "Drive the real CARLA server for a scenario, streaming frames to B2.\n\nReturns 503 (not 500) with an actionable message when CARLA is not installed\non this host or its server is unreachable \u2014 the guarded, documented handling\nfor the platform-restricted `carla` wheel (see service/carla_runner.py). The\nrun is synchronous; a production deployment would enqueue it as a job.",
        "operationId": "run_scenario_endpoint_scenarios__scenario_id__run_post",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeMetadata"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Scenario Endpoint",
        "tags": [
          "simulate"
        ]
      }
    },
...
  }
}